unsupervised-ml-api/requirements.txt
uvicorn>=0.20.0,<0.30.0
fastapi>=0.88.0,<1.0.0
python-multipart>=0.0.5,<0.1.0
pydantic>=1.10.4,<1.12.0
typing_extensions>=4.2.0,<5.0.0
loguru>=0.5.3,<1.0.0
scikit-learn>=1.1.3,<2.0.0
numpy>=1.21.0,<2.0.0
pandas>=1.3.5,<2.0.0
matplotlib
seaborn
unsupervised-model-pipe